handle not-checked-out submodules
If we have not updated a submodule with "git submodule update", we can
get the "(not checkout out)" message in git diff.

The simple split() assignment mechanism was raising a ValueError because
there were too many values to unpack. Use a regex instead to be more
precise, so we can capture the exact values we need.