squash everything into one .patch file
Squash all changes into a single .patch file named after the submodule.
For example, for src/civetweb, this generates a single "civetweb.patch"
file.

The reason we're doing this is to reduce the number of patches we're
having to manage in %prep. For example, a recent ceph change pulled in a
new version of civetweb that means tons of patches.